test: Skip tests on missing subsystems gracefully
All of the SDL2 subsystems are optional. Add an appropriate logic to
skip tests if SDL2 is built without them.

While missing audio and video subsystems are rather unlikely,
e.g. the haptic subsystem is rarely actually necessary (especially with
no compatible hardware). However, supporting every (or least most of)
missing subsystems should be cleaner and not cause any harm.

While technically SDL2_Init() and related functions can fail with other
errors, it is rather unlikely that the tests would be able to trigger
any real problem with the code that would result in non-zero return
value of this function.

Closes: https://github.com/marcusva/py-sdl2/issues/116