Set color key for indexed images
Calling load_image using a PNG file with indexed colors discarded the transparent palette index value, so all transparent pixels turned into whatever solid color is internally stored in the file that isn't typically seen. This change takes note of the transparency index and sets this as the color key on the resulting SDL surface, so that blitting operations will treat it as transparent.

The check for isinstance exists because info['transparency'] may not be an int: if the file is a PNG file with a per-palette alpha value it will be a bytearray, which is not supported here. See https://pillow.readthedocs.io/en/stable/handbook/image-file-formats.html for details on this.