test: Add more realistic tests of SDL_VERSIONNUM
The minor version overflows into the thousands digit in 2.23.0 and up.
Because the major version is fixed at 2 until SDL 3 (which will be a
separate library, like the difference between SDL 1.2 and 2), the
numbers from here can still be used as a sort order, but are now a bit
more confusing than they used to be.

Signed-off-by: Simon McVittie <smcv@collabora.com>