version_test: Don't assert that SDL_GetRevision() starts with http
The default format is going to change in 2.25.x, and in git main it
currently starts with "SDL-".

SDL specifically documents the result of this function as "not intended
to be reliable in any way", so it seems wrong to have pysdl2's tests fail
whenever the format changes. To address that, turn unexpected formats
into an xfail, so that they're flagged as something to investigate but
do not make the unit tests fail when used as a QA gate.

Closes: https://github.com/py-sdl/py-sdl2/issues/248
Signed-off-by: Simon McVittie <smcv@collabora.com>