tests: Skip pixel comparison when loading XCF files on big-endian
This isn't working for me. On x86 I get the desired rendering, but on
s390x, for example the pixels in the first row that should have been
green and red come out as transparent black.

There are no big-endian architectures that are relevant for gaming this
decade, and XCF is not a widespread format for game use, so it seems
reasonable to ignore this.

Signed-off-by: Simon McVittie <smcv@collabora.com>