Skip test for big-endian .xcf loading (#259)
* tests: Test image loading in a predictable order

This makes it easier to compare different architectures' ability to load
the various file formats.

Signed-off-by: Simon McVittie <smcv@collabora.com>

* ext.image: Make comments and variable naming less confusing

Somewhat confusingly, SDL uses ABGR8888 to represent a format where each
32-bit unit can be interpreted as 0xAABBGGRR in native endianness,
whereas it uses ARGB32 to represent a format where 8-bit units can be
interpreted as AA, RR, GG, BB in that order. Stick to the same
terminology as SDL here, and be clearer about what we mean.

No functional change intended.

Signed-off-by: Simon McVittie <smcv@collabora.com>

* tests: Skip pixel comparison when loading XCF files on big-endian

This isn't working for me. On x86 I get the desired rendering, but on
s390x, for example the pixels in the first row that should have been
green and red come out as transparent black.

There are no big-endian architectures that are relevant for gaming this
decade, and XCF is not a widespread format for game use, so it seems
reasonable to ignore this.

Signed-off-by: Simon McVittie <smcv@collabora.com>

---------

Signed-off-by: Simon McVittie <smcv@collabora.com>