tests: SDL_GetError() != b'' isn't an error
In Ubuntu, we're seeing test failures of the following pattern when
moving from libsdl2 2.28.1+dfsg-1 to 2.28.2+dfsg-1:

    > assert sdl2.SDL_GetError() == b""
    E AssertionError: assert b'Unexpected ...r element crc' == b''
    E   Full diff:
    E   - b''
    E   + b'Unexpected controller element crc'

This looks like an issue similar to that fixed in commit 8c39f40. We
should check the relevant return value, and only if it indicates failure
should we attach any particular meaning to the return value of
SDL_GetError().